Merges IAuthenticationService and IAuthenticationTokenProvider<AuthenticationData> into AuthenticationService implementation
